Split config entries
