+++ src/main/kotlin/cn/transfur/furbot/Config.kt
@@ -7,6 +7,19 @@ import net.mamoe.mirai.console.data.value
 import net.mamoe.yamlkt.Comment
 
 object Config : AutoSavePluginConfig("config") {
+    @ValueDescription("基本配置")
+    val global: Global by value()
+
+    @Serializable
+    data class Global(
+        @Comment("是否显示绒狸开源版尾巴（默认开启）")
+        val showTail: Boolean = true,
+        @Comment("是否响应私聊消息（默认关闭）")
+        val respondFriends: Boolean = false,
+        @Comment("是否响应群聊消息（默认开启）")
+        val respondGroups: Boolean = true
+    )
+
     @ValueDescription("与吸毛模块相关的配置")
     val furbot: Furbot by value()
 
@@ -15,13 +28,7 @@ object Config : AutoSavePluginConfig("config") {
         @Comment("申请开源版地址的 QQ 号码")
         val qq: Long = 0L,
         @Comment("申请开源版地址的授权码")
-        val authKey: String = "",
-        @Comment("是否显示绒狸开源版尾巴（默认开启）")
-        val showTail: Boolean = true,
-        @Comment("是否响应私聊消息（默认关闭）")
-        val respondFriends: Boolean = false,
-        @Comment("是否响应群聊消息（默认开启）")
-        val respondGroups: Boolean = true,
+        val authKey: String = ""
     )
 
     @ValueDescription(".晚安相关配置")
