+++ src/main/kotlin/cn/transfur/furbot/KotlinPluginMain.kt
@@ -2,10 +2,10 @@ package cn.transfur.furbot
 
 import cn.transfur.furbot.command.GroupOnlyCommand
 import cn.transfur.furbot.command.admin.FurbotControlCommand
-import cn.transfur.furbot.command.core.GetFidsByNameCommand
-import cn.transfur.furbot.command.core.GetFurByFidCommand
-import cn.transfur.furbot.command.core.GetFurByNameCommand
-import cn.transfur.furbot.command.core.GetFurRandCommand
+import cn.transfur.furbot.command.furbot.GetFidsByNameCommand
+import cn.transfur.furbot.command.furbot.GetFurByFidCommand
+import cn.transfur.furbot.command.furbot.GetFurByNameCommand
+import cn.transfur.furbot.command.furbot.GetFurRandCommand
 import cn.transfur.furbot.command.misc.GoodNightCommand
 import cn.transfur.furbot.command.misc.HelpCommand
 import cn.transfur.furbot.network.TailApiClient
