+++ src/main/kotlin/cn/transfur/furbot/command/furbot/GetFurByNameCommand.kt
@@ -1,4 +1,4 @@
-package cn.transfur.furbot.command.core
+package cn.transfur.furbot.command.furbot
 
 import cn.transfur.furbot.command.SessionCommand
 import cn.transfur.furbot.data.FurPic
